requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
asn1crypto==1.5.1
    # via scramp
beautifulsoup4==4.13.4
    # via redshift-connector
boto3==1.37.37
    # via redshift-connector
botocore==1.37.37
    # via
    #   boto3
    #   redshift-connector
    #   s3transfer
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via typer
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
google-api-core[grpc]==2.24.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.39.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
google-cloud-bigquery==3.31.0
    # via lead-recovery (pyproject.toml)
google-cloud-core==2.4.3
    # via google-cloud-bigquery
google-crc32c==1.7.1
    # via google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-bigquery
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jiter==0.9.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==5.3.2
    # via redshift-connector
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.5
    # via pandas
openai==1.75.0
    # via lead-recovery (pyproject.toml)
packaging==25.0
    # via
    #   google-cloud-bigquery
    #   redshift-connector
pandas==2.2.3
    # via lead-recovery (pyproject.toml)
proto-plus==1.26.1
    # via google-api-core
protobuf==5.29.4
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyarrow==19.0.1
    # via lead-recovery (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.3
    # via
    #   lead-recovery (pyproject.toml)
    #   openai
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.9.1
    # via lead-recovery (pyproject.toml)
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.1.0
    # via
    #   lead-recovery (pyproject.toml)
    #   pydantic-settings
pytz==2025.2
    # via
    #   pandas
    #   redshift-connector
redshift-connector==2.1.5
    # via lead-recovery (pyproject.toml)
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   redshift-connector
    #   tiktoken
rich==14.0.0
    # via typer
rsa==4.9.1
    # via google-auth
s3transfer==0.11.5
    # via boto3
scramp==1.4.5
    # via redshift-connector
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.6
    # via beautifulsoup4
tenacity==9.1.2
    # via lead-recovery (pyproject.toml)
tiktoken==0.9.0
    # via lead-recovery (pyproject.toml)
tqdm==4.67.1
    # via
    #   lead-recovery (pyproject.toml)
    #   openai
typer[all]==0.15.2
    # via lead-recovery (pyproject.toml)
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via
    #   botocore
    #   requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
psutil>=5.9.0  # For memory usage tracking
